+++ rules/node/bestPractices.js
@@ -1,0 +1,5 @@
+module.exports = {
+  rules: {
+    'node/no-deprecated-api': 'error',
+  },
+};