+++ rules/node/errors.js
@@ -1,0 +1,22 @@
+module.exports = {
+  rules: {
+    'node/handle-callback-err': 'off',
+    'node/no-callback-literal': 'off',
+    'node/no-exports-assign': 'error',
+    'node/no-extraneous-import': 'error',
+    'node/no-extraneous-require': 'error',
+    'node/no-missing-import': 'error',
+    'node/no-missing-require': 'error',
+    'node/no-new-require': 'error',
+    'node/no-path-concat': 'off',
+    'node/no-process-exit': 'error',
+    'node/no-unpublished-bin': 'error',
+    'node/no-unpublished-import': 'error',
+    'node/no-unpublished-require': 'error',
+    'node/no-unsupported-features/es-builtins': 'error',
+    'node/no-unsupported-features/es-syntax': 'error',
+    'node/no-unsupported-features/node-builtins': 'error',
+    'node/process-exit-as-throw': 'error',
+    'node/shebang': 'error',
+  },
+};